Graph Plotter (Spirographs)
Plots graph for particular functions. Best for plotting Spirographs.
Just put values of radius of outer and inner circles and the distance of
point whose path is to be traced 'd' w.r.t the center of inner circle.
Run the app-debug.apk on your android device to see the working.